Restore navigation functionality to the admin panel for improved usability
Reverts changes and restores the original admin panel page from a backup file (admin-panel-page-backup.tsx).
